This ActionScript allows the caller to change sssd debug_level for all nodes in a cluster.
This would allow a customer to better monitor sssd LDAP or AD
login issues when they arise, as the debug output would be
emitted by the sssd services.

The script edits the /etc/sssd/sssd.conf file, then restarts
the sssd service on the node so it may reinitialize using the
new debug level setting.

Signed-off-by: Anne Marie Merritt <amm@bluedata.com>