Sample pipelines for kubeflow
mnist_train:

 Dockerfile for creating the image and model file to deploy the image with the DOckerfile

mnist_test:

Dockerfile to create the image and model file to load the model and calcuate the accuracy metrics

Note: The Dockerfiles are not enabled with proxy. As a result, they shall not be able to download. Please set the empty proxy variables inside the dockerfiles to enable proxy.